requirements.txt
flask
flask_wtf
requests
python-dotenv
supabase
flaskwebgui
